requirements.txt
yt-dlp==2024.1.0
supabase==2.3.0
openai==1.12.0
watchdog==3.0.0
schedule==1.2.0
python-dotenv==1.0.0
requests==2.31.0
ffmpeg-python==0.2.0
apscheduler==3.10.4